pyproject.toml
[project]
name = "YetAnotherPicSearch4Telegram"
version = "1.0.0"
description = "Yet Another Picture Search bot for Telegram"
authors = [{ name = "NekoAria" }]
license = { text = "GPL-3.0-only" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "arrow>=1.3.0",
    "cachetools>=6.0.0",
    "httpx[http2]>=0.28.1",
    "imageio>=2.37.0",
    "loguru>=0.7.3",
    "lxml>=5.4.0",
    "PicImageSearch>=3.12.8",
    "Pillow>=11.2.1",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pyquery>=2.0.1",
    "Telethon>=1.40.0",
    "tenacity>=9.1.2",
]
keywords = [
    "telegram",
    "animetrace",
    "ascii2d",
    "baidu",
    "bing",
    "copyseeker",
    "e-hentai",
    "google",
    "google-lens",
    "iqdb",
    "saucenao",
    "tineye",
    "tracemoe",
    "yandex",
    "anime",
    "danbooru",
    "doujin",
    "pixiv",
]

[project.urls]
homepage = "https://github.com/NekoAria/YetAnotherPicSearch4Telegram"
repository = "https://github.com/NekoAria/YetAnotherPicSearch4Telegram"

[project.optional-dependencies]
socks = [
    "python-socks[asyncio]>=2.7.1",
    "socksio>=1.0.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.29.2",
    "pre-commit>=4.2.0",
    "ruff>=0.11.13",
]

[tool.basedpyright]
pythonVersion = "3.9"
reportAny = false
reportExplicitAny = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportUnknownVariableType = false

[tool.ruff]
target-version = "py39"
line-length = 120
extend-exclude = ["typings"]

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"*.py" = ["F405", "N813"]

[tool.ruff.lint.pydocstyle]
convention = "google"
